Split out the OpenMAX DL timing routines to their own file.
Update gyp file appropriately.
